# docs/api.yaml
openapi: 3.0.3

info:
  title: Meditation GitHub Stats API
  description: |
    A serverless API that generates embeddable SVG cards for GitHub profile statistics.
    All endpoints return cached SVG images suitable for embedding in GitHub READMEs
    or any web page via `<img>` tags.

    ## Embedding Example
    ```markdown
    ![Activity Graph](https://meditation-cyan.vercel.app/graph?user=fawwazgoreng&theme=winter)
    ![Top Languages](https://meditation-cyan.vercel.app/stats?user=fawwazgoreng&theme=dark)
    ![Streak](https://meditation-cyan.vercel.app/streak?user=fawwazgoreng)
    ```

    ## Caching
    All SVG responses are cached for **1 hour** (`Cache-Control: public, max-age=3600`).
    The cache key is composed of `endpoint + username + theme`.

  version: 1.1.0
  contact:
    name: fawwazgoreng
    url: https://github.com/fawwazgoreng

servers:
  - url: https://meditation-cyan.vercel.app
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local development

# ─── Reusable Components ──────────────────────────────────────────────────────

components:
  parameters:
    UserParam:
      name: user
      in: query
      required: true
      description: GitHub username to fetch data for.
      schema:
        type: string
        example: fawwazgoreng

    ThemeParam:
      name: theme
      in: query
      required: false
      description: |
        Sets the color scheme for the SVG card. If omitted, it defaults to the `dark` theme.

        **Theme Catalog:**
        * `dark` - Standard GitHub Dark with ocean blue accents.
        * `light` - Classic, clean GitHub Light style.
        * `pickmeup` - Dark game UI — mana blue title, gacha red accent.
        * `northernblade` -  Pitch black ink — silent night purple.
        * `orv` -  Omniscient Reader — gold title, transcendence blue accent.
        * `winter` - Calm ice-blue and navy tones.
        * `ocean` - Deep sea blue with bright text.
        * `cyberpunk` - High contrast featuring black, fuchsia, and cyan.
        * `galaxy` - Space-themed purple gradients with star accents.
        * `forest` - Dark emerald green for a natural feel.
        * `volcano` - Deep black with fiery red accents.
        * `sunset` - Dark rose gradients with warm red tones.
        * `zen` - Cream white with minimalist charcoal grey.
      schema:
        type: string
        default: dark
        example: winter
        enum:
          - dark
          - light
          - winter
          - ocean
          - cyberpunk
          - galaxy
          - forest
          - volcano
          - sunset
          - zen

  responses:
    SvgResponse:
      description: SVG image generated successfully.
      headers:
        Cache-Control:
          schema:
            type: string
            example: public, max-age=3600, s-maxage=3600
        Content-Type:
          schema:
            type: string
            example: image/svg+xml
      content:
        image/svg+xml:
          schema:
            type: string

    MissingUserError:
      description: Required `user` parameter is missing.
      content:
        text/plain:
          schema:
            type: string
            example: "Error: query param 'user' is required."

    UpstreamError:
      description: GitHub API error or user not found.
      content:
        text/plain:
          schema:
            type: string
            example: "Error: User not found"

# ─── Tags ────────────────────────────────────────────────────────────────────

tags:
  - name: Health
    description: Service availability and status checks.
  - name: Stats
    description: GitHub profile statistics rendered as SVG cards.

# ─── Paths ───────────────────────────────────────────────────────────────────

paths:
  # ── Root ──────────────────────────────────────────────────────────────────

  /:
    get:
      tags: [Health]
      summary: Service status
      operationId: getStatus
      responses:
        "200":
          description: Service is active.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "active" }
                  endpoints:
                    type: array
                    items: { type: string }
                    example:
                      [
                        "/stats",
                        "/streak",
                        "/card",
                        "/graph",
                      ]

  # ── Contribution Graph (Line Chart) ────────────────────────────────────────

  /graph:
    get:
      tags: [Stats]
      summary: 30-Day Contribution Activity Graph
      description: |
        Generates a professional line chart representing the user's contribution activity over the **last 30 days**.

        **Features:**
        - **Line Chart & Area Fill:** Visual representation of daily commits with a smooth area fill.
        - **Daily Date Labels:** Rotated X-axis labels for every day of the month.
        - **Contribution Axis:** Y-axis indicating specific contribution counts.
        - **Peak Markers:** Visual indicators for the most active days.
        - **Grid System:** Dashed horizontal and vertical lines for enhanced readability.

        **SVG dimensions:** 600 × 320px
        **Cache key:** `graph:{user}:{theme}`
      operationId: getActivityGraph
      parameters:
        - $ref: "#/components/parameters/UserParam"
        - $ref: "#/components/parameters/ThemeParam"
      responses:
        "200":
          $ref: "#/components/responses/SvgResponse"
        "400":
          $ref: "#/components/responses/MissingUserError"
        "502":
          $ref: "#/components/responses/UpstreamError"

  # ── Top Languages ──────────────────────────────────────────────────────────

  /stats:
    get:
      tags: [Stats]
      summary: Top programming languages card
      description: Returns a breakdown of the most used programming languages in the user's public repositories.
      operationId: getTopLanguages
      parameters:
        - $ref: "#/components/parameters/UserParam"
        - $ref: "#/components/parameters/ThemeParam"
      responses:
        "200":
          $ref: "#/components/responses/SvgResponse"
        "400":
          $ref: "#/components/responses/MissingUserError"
        "502":
          $ref: "#/components/responses/UpstreamError"

  # ── Streak ─────────────────────────────────────────────────────────────────

  /streak:
    get:
      tags: [Stats]
      summary: Contribution streak card
      description: Displays the current streak, longest streak, and total contributions for the current year.
      operationId: getStreak
      parameters:
        - $ref: "#/components/parameters/UserParam"
        - $ref: "#/components/parameters/ThemeParam"
      responses:
        "200":
          $ref: "#/components/responses/SvgResponse"
        "400":
          $ref: "#/components/responses/MissingUserError"
        "502":
          $ref: "#/components/responses/UpstreamError"

  # ── Stats Card ─────────────────────────────────────────────────────────────

  /card:
    get:
      tags: [Stats]
      summary: General profile statistics card
      description: Provides an overview of general GitHub profile statistics such as total stars, followers, and public repos.
      operationId: getStatsCard
      parameters:
        - $ref: "#/components/parameters/UserParam"
        - $ref: "#/components/parameters/ThemeParam"
      responses:
        "200":
          $ref: "#/components/responses/SvgResponse"
        "400":
          $ref: "#/components/responses/MissingUserError"
        "502":
          $ref: "#/components/responses/UpstreamError"